pyproject.toml
[project]
name = "darp"
version = "0.1.0"
description = "Darp "
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "curl-cffi>=0.7.4",
    "httpx>=0.28.1",
    "mcp>=1.2.0",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "web3>=7.6.1",
]
